+++ packages/author-profile/__tests__/android/author-profile.test.js
@@ -1,5 +1,5 @@
-import shared from "../shared";
 import sharedNative from "../shared-native";
+import shared from "../author-profile";
 
 describe("AuthorProfile tests on android", () => {
   shared();