+++ packages/topic/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.3"></a>
+## [0.2.3](https://github.com/newsuk/times-components/compare/@times-components/topic@0.2.2...@times-components/topic@0.2.3) (2018-05-25)
+
+
+
+
+**Note:** Version bump only for package @times-components/topic
+
 <a name="0.2.2"></a>
 ## [0.2.2](https://github.com/newsuk/times-components/compare/@times-components/topic@0.2.1...@times-components/topic@0.2.2) (2018-05-23)