+++ packages/video/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.3.4"></a>
+## [1.3.4](https://github.com/newsuk/times-components/compare/@times-components/video@1.3.3...@times-components/video@1.3.4) (2018-05-25)
+
+
+### Bug Fixes
+
+* update react-native-svg for expo ([#959](https://github.com/newsuk/times-components/issues/959)) ([0dbef64](https://github.com/newsuk/times-components/commit/0dbef64))
+
+
+
+
 <a name="1.3.3"></a>
 ## [1.3.3](https://github.com/newsuk/times-components/compare/@times-components/video@1.3.2...@times-components/video@1.3.3) (2018-05-23)